refactor: create an index to all utils
