--- src/utils/categories.js
@@ -1,5 +1,0 @@
-import { themes } from '../styles/themes';
-
-export const getCategoryBg = (category) => {
-  return themes.palette.categories[category];
-};
